# specs/openapi-v2-complete.yaml
openapi: "3.1.0"
info:
  title: IPED Web API — complete v2
  description: |
    Full REST contract for iped-webapi (Jersey/JAX-RS on embedded Jetty)
    covering Phases 1–5 of the distributed-processing roadmap.

    Phase 1 endpoints (`/cases/{caseId}/…`) are retained for reference.
    All new functionality is under `/v2/…` paths.

    Authentication: `X-Api-Key` header or `Authorization: Bearer <key>`.
    When no API key is configured the server runs in open mode.
  version: "0.5.0"
  contact:
    name: IPED Project
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:39000
    description: Local engine (default port)
security:
  - ApiKeyHeader: []
  - BearerToken: []

# ─────────────────────────────────────────────────────────────────────────────
# Tags
# ─────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /v2/jobs:
    get:
      tags: [jobs]
      operationId: listJobs
      summary: List all active and recent jobs
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: "#/components/schemas/JobStatus"
    post:
      tags: [jobs]
      operationId: createJob
      summary: Submit an async job (export or report)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
  /v2/jobs/{jobId}:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [jobs]
      operationId: getJob
      summary: Get job status
      responses:
        "200":
          description: Current job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [jobs]
      operationId: cancelJob
      summary: Cancel a running job
      responses:
        "204":
          description: Job cancelled (or already terminal)
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/audit/{caseId}:
    parameters:
      - $ref: "#/components/parameters/CaseId"
    get:
      tags: [audit]
      operationId: getCaseAudit
      summary: All terminal processing records for a case (newest first)
      responses:
        "200":
          description: Audit records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditPage"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/audit/{caseId}/csv:
    parameters:
      - $ref: "#/components/parameters/CaseId"
    get:
      tags: [audit]
      operationId: getCaseAuditCsv
      summary: All records as RFC 4180 CSV download
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/audit/{caseId}/{itemUuid}:
    parameters:
      - $ref: "#/components/parameters/CaseId"
      - name: itemUuid
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [audit]
      operationId: getItemAudit
      summary: Most recent terminal record for one item
      responses:
        "200":
          description: Processing record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessingRecord"
        "404":
          $ref: "#/components/responses/NotFound"

# ─────────────────────────────────────────────────────────────────────────────
# Components
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-Api-Key
    BearerToken:
      type: http
      scheme: bearer
  parameters:
    CaseId:
      name: caseId
      in: path
      required: true
      schema:
        type: string
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
...
    CreateJobRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [export, report]
        scope:
          type: string
          enum: [checked, all]
          default: checked
        format:
          type: string
          enum: [zip, pdf, html]
          default: zip
        caseId:
          type: string
    JobStatus:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        type:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
...
    AuditPage:
      type: object
      properties:
        caseId:
          type: string
        total:
          type: integer
        records:
          type: array
          items:
            $ref: "#/components/schemas/ProcessingRecord"
    ProcessingRecord:
      type: object
      properties:
        itemUuid:
          type: string
        caseId:
          type: string
        taskType:
          type: string
        pipelineStage:
          type: integer
        agentId:
          type: string
        processedAt:
          type: string
          format: date-time
        durationMs:
          type: integer
          format: int64
        outcome:
          type: string
          enum: [COMPLETED, ERROR, TIMEOUT]
        errorMessage:
          type: string